+++ creature.py
@@ -1,9 +1,8 @@
 import random
 import pygame
 import torch
-import torch.nn as nn
+from torch import nn
 import torch.nn.functional as F
-import cows
 
 
 class Creature:
@@ -11,18 +10,15 @@ class Creature:
     _photosynthesis_scale = 0.01
     _maximum_age = 2001
 
-    def __init__(self, x: int, y: int, screen: pygame.Surface, image: str = "images/blank.png"):
-        self.screen = screen
+    def __init__(self, x: int, y: int, image: str = "images/blank.png"):
         self.x = x
         self.y = y
-        self.im: pygame.Surface = pygame.image.load(image)
-        self.rect = self.im.get_rect()
-        self.rect.move(self.x, self.y)
+        self.image: pygame.Surface = pygame.image.load(image)
+        rect = self.image.get_rect()
+        rect.move(self.x, self.y)
         # Physical attributes
-        self.dead = False
         self.age = 0
         self.mass = 0
-        self.reproduction_tick = Creature._default_reproduction_tick
         # 0 - speed
         # 1 - photosynthesis
         # 2 - reproduction: 0 - asexual; 1 - sexual
@@ -43,19 +39,18 @@ class Creature:
     def set_sexual_reproduction(self, reproduction: int):
         self.genome[2] = reproduction
 
-    def draw(self):
+    def draw(self, game_state):
         # colors:
         photosynthesis = self.get_photosynthesis()
         greenness = photosynthesis * 255 if photosynthesis > 0 else 0
         blueness = 255 if photosynthesis == 0 else 0
-        if greenness > 255:
-            greenness = 255
-        image = self.im.copy()
+        greenness = min(greenness, 255)
+        image = self.image.copy()
         image = pygame.transform.scale(image, (self.mass, self.mass))
         image.fill((0, greenness, blueness, 255), special_flags=pygame.BLEND_RGBA_ADD)
-        self.screen.blit(image, (self.x, self.y))
+        game_state.screen.blit(image, (self.x, self.y))
 
-    def logic(self, game_state: cows.Game):
+    def logic(self, game_state):
         # Behave
         self.reproduce(game_state)
         self.tick_age()
@@ -69,37 +64,30 @@ class Creature:
     def photosynthesise(self):
         self.mass = self.mass + self.get_photosynthesis() * Creature._photosynthesis_scale
 
-    def reproduce(self, game: cows.Game):
+    def reproduce(self, game_state):
         if self.genome[2] is not None:
             if not self.is_sexual_reproduction():
-                if self.reproduction_tick <= 0:
+                if self.age != 0 and self.age % Creature._default_reproduction_tick == 0:
                     r_x = random.randint(-50, 50)
                     r_y = random.randint(-50, 50)
                     new_grass = Creature(
-                        self.x + r_x, self.y + r_y, game.screen)
+                        self.x + r_x, self.y + r_y)
                     new_grass.genome = self.genome.copy()
-                    game.add_creature(new_grass)
-                    self.reset_reproduction_tick()
-                self.reproduction_tick -= 1
+                    game_state.add_creature(new_grass)
 
     def is_sexual_reproduction(self):
         return self.genome[2] == 1
 
-    def reset_reproduction_tick(self):
-        self.reproduction_tick = Creature._default_reproduction_tick
-
     def tick_age(self):
         if self.age >= Creature._maximum_age:
-            self.dead = True
             return
         self.age += 1
 
     def is_dead(self):
-        return self.dead
+        return self.age > self._maximum_age
 
     def think(self, vision):
-        foobar = self.net(vision)
-        print(foobar)
+        pass
 
     def get_vision(self):
         return self.genome[3]
